layer.conf: Update layer dependencies
Specifically list all of the layer dependencies.

Circular layer dependencies are allowed in OE, however the layerindexer can
not handle this.  Otherwise it would be a good idea to have overc and cube
both depend on each.  This dependency was broken in favor of overc depending
on cube, as cube will not do anything with an overc distribution.

The list of dependencies was based on the conf/bblayers.conf.sample file.

cloud-services-layer is required for consul

(LOCAL REV: NOT UPSTREAM) -- TBD

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>

Issue: LINUXEXEC-1875
Rebase to wr-10.19-20190520
Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>