packagegroups: Change from PACKAGE_ARCH to MACHINE_ARCH
Many of these groups have indirect checks to see if grub is available, based
on the MACHINE_FEATURES.  This can change from BSP to BSP, and thus the hash
and PR values for these recipes change.

Instead of permitting this indeterminite behavior, we want to make these
select packagegroups to be MACHINE_ARCH, so they will be unique to the needs
of the specific BSP they are building against.

(LOCAL REV: NOT UPSTREAM) -- Pull request sent to github #248

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>