pflask: switch from clone flag CLONE_UNCONTAIN to CAP_ADMIN_NS capability bits
In the 5.2 kernel and beyond the clone() flags no longer contain any
space to add the CLONE_UNCONTAIN flag.  The kernel was changed to only
pass on the CAP_ADMIN_NS to a level 1 container when the CAP_ADMIN_NS
ambient and inheritable flags are set.

The arguments to pflask remain the same as before for setting up the
uncontain capabilities.

[ Issue: LIN1019-2544 ]

(LOCAL REV: NOT UPSTREAM) -- Sent https://github.com/OverC/meta-overc/pull/272

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>