overc-utils: dom0 needs to depend on util-linux-script
The "c3-console dom0" from cube-server no longer works due to the
further split of the utilities in the oe-core of util-linux.  The
simple fix is to just add util-linux-script as a dependency to the
dom0 packagegroup.

[ Issue: LIN1019-2953 ]

(master rev: 7a4e42e9ff3a45a7859f45ec2b2443ab56cccd84)

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>